変更内容：ファイル分割
理由：メインコードが長くなりすぎて、可読性が下がったため
