+++ 04.ls/lib/my-ls.rb
@@ -4,46 +4,10 @@
 require 'optparse'
 require 'etc'
 
+require_relative 'jp_methods.rb'
+require_relative 'conversion_file_info_methods.rb'
+
 LIST_ROW_NUM = 3
-
-def divide_equal(file_names, number)
-  split_num = (file_names.size / number.to_f).ceil
-  file_names.each_slice(split_num).to_a
-end
-
-def transpose_lack(uneven_size_array)
-  # Array#teansposeとは異なり、二次元配列の要素のサイズが異なっても転置できるメソッド
-  max_size = uneven_size_array.map(&:size).max
-  uneven_size_array.map { |element| element + [nil] * (max_size - element.size) }.transpose.map(&:compact)
-end
-
-def sort_jp(jp_array)
-  # Array#sortとは異なり、漢字→ひらがな→カタカナの順にソートするメソッド
-  jp_array.sort do |a, b|
-    if a.match?(/\p{Han}/) && b.match?(/\p{Hiragana}|\p{Katakana}/)
-      -1
-    elsif a.match?(/\p{Hiragana}|\p{Katakana}/) && b.match?(/\p{Han}/)
-      1
-    else
-      a <=> b
-    end
-  end
-end
-
-def size_jp(jp_string)
-  # String#sizeと異なり、日本語を2文字とみなすメソッド
-  jp_string.each_char.sum do |char|
-    if char.match?(/\p{Han}|\p{Hiragana}|\p{Katakana}|ー|（|）/)
-      2
-    else
-      1
-    end
-  end
-end
-
-def ljust_jp(jp_string, max_size)
-  jp_string + ' ' * (max_size - size_jp(jp_string))
-end
 
 def get_file_names(argument_name, options)
   argument_name ||= '.'
@@ -74,13 +38,19 @@ def select_files(target_dir, target_file, options)
   else
     file_names_all.select! { |file_name| file_name == target_file } # '.ファイル名'も表示対象
   end
-  options['l'] ? get_files_info_merged_text(target_dir, file_names_all) : file_names_all
+  options['l'] ? get_files_info_text_merged(target_dir, file_names_all) : file_names_all
 end
 
-def get_files_info_merged_text(target_dir, file_names_all)
+def get_files_info_text_merged(target_dir, file_names_all)
   files_info_each_type = get_files_info_each_type(target_dir, file_names_all)
   files_info_text = files_info_each_type['file_name'].map.with_index do |_name, idx|
-    "#{files_info_each_type['mode'][idx]} #{files_info_each_type['number_of_link'][idx]} #{files_info_each_type['user_name'][idx]}  #{files_info_each_type['group_name'][idx]}  #{files_info_each_type['size'][idx]}  #{files_info_each_type['mtime'][idx]} #{files_info_each_type['file_name'][idx]}"
+    "#{files_info_each_type['mode'][idx]} "\
+    "#{files_info_each_type['number_of_link'][idx]} "\
+    "#{files_info_each_type['user_name'][idx]}  "\
+    "#{files_info_each_type['group_name'][idx]}  "\
+    "#{files_info_each_type['size'][idx]}  "\
+    "#{files_info_each_type['mtime'][idx]} "\
+    "#{files_info_each_type['file_name'][idx]}"
   end
   ["total #{file_names_all.map { |file_name| File::Stat.new("#{target_dir}/#{file_name}") }.map(&:blocks).sum}"] + files_info_text
 end
@@ -98,58 +68,6 @@ def get_files_info_each_type(target_dir, file_names_all)
   files_info_each_type
 end
 
-def convert_files_mode_to_l_option_format(files_mode)
-  files_mode_chars = files_mode.map do |file_mode|
-    file_mode_bits = format('%016b', file_mode)
-    {
-      'file_type' => convert_file_type_bit_to_char(file_mode_bits[0..3]),
-      'owner_permission' => convert_permission_bits_to_str(file_mode_bits[7..9], file_mode_bits[4]),
-      'group_permission' => convert_permission_bits_to_str(file_mode_bits[10..12], file_mode_bits[5]),
-      'others_permission' => convert_permission_bits_to_str(file_mode_bits[13..15], file_mode_bits[6])
-    }
-  end
-  files_mode_chars.map { |chars| chars['file_type'] + chars['owner_permission'] + chars['group_permission'] + chars['others_permission'] + '@' }
-end
-
-def convert_file_type_bit_to_char(file_type_bit)
-  file_type_char = { # file_modeに対する、8進数対応表 cf. Linux file type and mode Doc
-    '0o01'.to_i(8) =>	'p', # FIFO
-    '0o02'.to_i(8) => 'c', # Character special file
-    '0o04'.to_i(8) =>	'd', # Directory(ディレクトリ)
-    '0o06'.to_i(8) =>	'b', # Block special file
-    '0o10'.to_i(8) =>	'-', # Regular file(通常ファイル)
-    '0o12'.to_i(8) =>	'l', # Symbolic link(シンボリックリンク)
-    '0o14'.to_i(8) =>	's'	 # Socket link
-  }
-  file_type_char[file_type_bit.to_i(2)]
-end
-
-def convert_permission_bits_to_str(permission_bits, special_permission_bit)
-  file_permission = ['-', '-', '-']
-  file_permission[0] = 'r' if permission_bits[0] == '1'
-  file_permission[1] = 'w' if permission_bits[1] == '1'
-  file_permission[2] = ['-x', 'Ss'][special_permission_bit.to_i][permission_bits[2].to_i] # 2次元テーブルから実行権限記号を選択
-  file_permission.join
-end
-
-def convert_files_mtime_to_l_option_format(files_mtime)
-  files_each_mtime = {}
-  files_each_mtime['month'] = align_str_list_to_right(files_mtime.map(&:month).map(&:to_s))
-  files_each_mtime['day'] = align_str_list_to_right(files_mtime.map(&:day).map(&:to_s))
-  files_each_mtime['time'] = align_str_list_to_right(files_mtime.map { |mtime| "#{mtime.hour}:#{mtime.min}" })
-  [files_each_mtime['month'], files_each_mtime['day'], files_each_mtime['time']].transpose.map { |each_mtime| each_mtime.join(' ') }
-end
-
-def align_str_list_to_right(str_list)
-  max_size_str = str_list.map(&:size).max
-  str_list.map { |str| str.rjust(max_size_str) }
-end
-
-def align_jp_str_list_to_left(str_list)
-  max_size_str = str_list.map { |str| size_jp(str) }.max
-  str_list.map { |str| ljust_jp(str, max_size_str) }
-end
-
 def generate_name_list_text(file_names, number, options)
   row_max_num = options['l'] ? 1 : number
   separatiopn_names = divide_equal(file_names, row_max_num)
@@ -158,6 +76,17 @@ def generate_name_list_text(file_names, number, options)
   transpose_lack(separatiopn_names).inject('') do |text, names|
     text + "#{names[0..-2].map { |name| "#{ljust_jp(name, max_name_size)} " }.join}#{names[-1]}\n"
   end
+end
+
+def divide_equal(file_names, number)
+  split_num = (file_names.size / number.to_f).ceil
+  file_names.each_slice(split_num).to_a
+end
+
+def transpose_lack(uneven_size_array)
+  # Array#teansposeとは異なり、二次元配列の要素のサイズが異なっても転置できるメソッド
+  max_size = uneven_size_array.map(&:size).max
+  uneven_size_array.map { |element| element + [nil] * (max_size - element.size) }.transpose.map(&:compact)
 end
 
 options = {}
